pyproject.toml
[project]
name = "company-enricher"
version = "0.1.0"
description = "Async pipeline to enrich UK companies with websites, headcount & geo"
authors = [{ name = "Adam Rudd", email = "adam@mather.capital" }]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["companies", "enrichment", "async", "uk", "business-data"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "httpx>=0.27.0",
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "polars[lazy]>=0.20.0",
    "aiofiles>=23.2.0",
    "diskcache>=5.6.0",
    "duckduckgo-search>=5.3.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "stream-read-xbrl>=0.0.46",
    "pdfminer.six>=20221105",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "openpyxl>=3.1.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
company-enricher = "company_enricher.cli:app"

[project.urls]
Homepage = "https://github.com/your-org/company-enricher"
Repository = "https://github.com/your-org/company-enricher"
Issues = "https://github.com/your-org/company-enricher/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/company_enricher",
    "--cov-report=term-missing",
    "--cov-report=html",
]